tests/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
backports.entry-points-selectable==1.1.0
    # via virtualenv
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via pynacl
charset-normalizer==2.0.3
    # via requests
coverage==5.5
    # via -r requirements.in
deprecated==1.2.12
    # via pygithub
distlib==0.3.2
    # via virtualenv
dj-database-url==0.5.0
    # via django-app-helper
django==2.2.24
    # via
    #   -r requirements.in
    #   django-classy-tags
    #   django-cms
    #   django-extensions
    #   django-formtools
    #   django-sekizai
    #   django-treebeard
django-app-helper[cms]==3.0.1
    # via -r requirements.in
django-classy-tags==2.0.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==3.7.4
    # via django-app-helper
django-extensions==3.1.3
    # via -r requirements.in
django-formtools==2.3
    # via django-cms
django-sekizai==2.0.0
    # via django-cms
django-treebeard==4.5.1
    # via django-cms
djangocms-admin-style==2.0.2
    # via django-cms
docopt==0.6.2
    # via django-app-helper
filelock==3.0.12
    # via
    #   tox
    #   virtualenv
flake8==3.9.2
    # via -r requirements.in
idna==3.2
    # via requests
isort==5.9.2
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
mock==4.0.3
    # via -r requirements.in
packaging==21.0
    # via tox
platformdirs==2.1.0
    # via virtualenv
pluggy==0.13.1
    # via tox
py==1.10.0
    # via tox
pycodestyle==2.7.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.3.1
    # via
    #   -r requirements.in
    #   flake8
pygithub==1.55
    # via -r requirements.in
pyjwt==2.1.0
    # via pygithub
pynacl==1.4.0
    # via pygithub
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2021.1
    # via django
requests==2.26.0
    # via
    #   -r requirements.in
    #   pygithub
six==1.16.0
    # via
    #   django-app-helper
    #   django-cms
    #   pynacl
    #   python-dateutil
    #   tox
    #   virtualenv
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via tox
tox==3.24.0
    # via -r requirements.in
urllib3==1.26.6
    # via requests
virtualenv==20.6.0
    # via tox
wrapt==1.12.1
    # via deprecated
